# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "這是一個使用 Gin 框架構建的 CRUD API 服務，包含用戶管理、書籍管理、產品管理、文章管理、認證、限流器、工作池等功能。",
        "title": "Go Gin CRUD API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/rate-limiter/config": {
            "post": {
                "description": "為指定 key 設置限流器配置",
                "tags": [
                    "rate-limiter"
                ],
                "summary": "設置限流器配置",
                "parameters": [
                    {
                        "description": "限流器配置",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.RateLimiterConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.RateLimiterStatus"
                        }
                    }
                }
            }
        },
        "/api/rate-limiter/reset": {
            "post": {
                "description": "重置指定 key 的限流器",
                "tags": [
                    "rate-limiter"
                ],
                "summary": "重置限流器",
                "parameters": [
                    {
                        "type": "string",
                        "description": "限流鍵",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/rate-limiter/status": {
            "get": {
                "description": "獲取指定 key 的限流器狀態",
                "tags": [
                    "rate-limiter"
                ],
                "summary": "獲取限流器狀態",
                "parameters": [
                    {
                        "type": "string",
                        "description": "限流鍵（如：IP、用戶ID等）",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.RateLimiterStatus"
                        }
                    }
                }
            }
        },
        "/api/rate-limiter/test": {
            "post": {
                "description": "併發測試限流器的功能",
                "tags": [
                    "rate-limiter"
                ],
                "summary": "測試限流器",
                "parameters": [
                    {
                        "description": "測試請求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.RateLimiterTestRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.RateLimiterTestResponse"
                        }
                    }
                }
            }
        },
        "/api/rate-limiter/test/allow": {
            "get": {
                "description": "測試指定 key 的單個請求是否允許",
                "tags": [
                    "rate-limiter"
                ],
                "summary": "測試單個請求",
                "parameters": [
                    {
                        "type": "string",
                        "description": "限流鍵",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.RateLimiterStatus"
                        }
                    }
                }
            }
        },
...
        "/products": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "獲取產品列表，支援分頁和搜尋（需要認證）",
                "tags": [
                    "product"
                ],
                "summary": "獲取產品列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "頁碼",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每頁數量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜尋關鍵字",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.PaginationResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "創建一個新產品（需要管理員權限）",
                "tags": [
                    "product"
                ],
                "summary": "創建產品",
                "parameters": [
                    {
                        "description": "產品資訊",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根據 ID 獲取產品詳細資訊（需要認證）",
                "tags": [
                    "product"
                ],
                "summary": "獲取單一產品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "產品 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新指定產品的資訊（需要管理員權限）",
                "tags": [
                    "product"
                ],
                "summary": "更新產品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "產品 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新資訊",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-gin-crud_internal_dto.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "刪除指定產品（需要管理員權限）",
                "tags": [
                    "product"
                ],
                "summary": "刪除產品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "產品 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-gin-crud_internal_dto.CreateProductRequest": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
        "go-gin-crud_internal_dto.PaginationResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "go-gin-crud_internal_dto.ProductResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "go-gin-crud_internal_dto.RateLimiterConfig": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "description": "演算法類型：sliding_window, fixed_window",
                    "type": "string"
                },
                "key": {
                    "description": "限流鍵（如：IP、用戶ID等）",
                    "type": "string"
                },
                "limit": {
                    "description": "限制數量",
                    "type": "integer"
                },
                "window": {
                    "description": "時間窗口（如：1s, 1m）",
                    "type": "string"
                }
            }
        },
...
        "go-gin-crud_internal_dto.RateLimiterStatus": {
            "type": "object",
            "properties": {
                "algorithm": {
                    "description": "演算法類型",
                    "type": "string"
                },
                "current_count": {
                    "description": "當前窗口內的請求數",
                    "type": "integer"
                },
                "is_allowed": {
                    "description": "是否允許請求",
                    "type": "boolean"
                },
                "key": {
                    "type": "string"
                },
                "limit": {
                    "type": "integer"
                },
                "remaining": {
                    "description": "剩餘可用請求數",
                    "type": "integer"
                },
                "reset_time": {
                    "description": "重置時間",
                    "type": "string"
                },
                "window": {
                    "description": "時間窗口字串表示",
                    "type": "string"
                }
            }
        },
        "go-gin-crud_internal_dto.RateLimiterTestRequest": {
            "type": "object",
            "required": [
                "key",
                "limit",
                "requests",
                "window"
            ],
            "properties": {
                "concurrent": {
                    "description": "併發數（可選）",
                    "type": "integer"
                },
                "key": {
                    "description": "限流鍵",
                    "type": "string"
                },
                "limit": {
                    "description": "限制數量",
                    "type": "integer",
                    "minimum": 1
                },
                "requests": {
                    "description": "測試請求數量",
                    "type": "integer",
                    "minimum": 1
                },
                "window": {
                    "description": "時間窗口（如：1s, 1m）",
                    "type": "string"
                }
            }
        },
        "go-gin-crud_internal_dto.RateLimiterTestResponse": {
            "type": "object",
            "properties": {
                "allowed_requests": {
                    "type": "integer"
                },
                "blocked_requests": {
                    "type": "integer"
                },
                "details": {
                    "description": "詳細結果",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-gin-crud_internal_dto.RateLimiterStatus"
                    }
                },
                "duration": {
                    "description": "測試耗時",
                    "type": "string"
                },
                "success_rate": {
                    "description": "成功率",
                    "type": "number"
                },
                "total_requests": {
                    "type": "integer"
                }
            }
        },
...
        "go-gin-crud_internal_dto.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "請在值前加上 \"Bearer \" 前綴，例如：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}